+++ tests/wdio/beAccessible-spec.js
@@ -1,9 +1,13 @@
-describe('axe', () => {
+describe('be accessible', () => {
   const viewports = Terra.viewports('tiny', 'huge');
 
   describe('accessible', () => {
     before(() => browser.url('/accessible.html'));
     Terra.should.beAccessible({ viewports });
+    Terra.it.isAccessible({ viewports });
+    it('is accessible', () => {
+      Terra.validates.accessibility({ viewports });
+    });
   });
 
   describe('inaccessible contrast', () => {
@@ -13,5 +17,9 @@ describe('axe', () => {
 
     before(() => browser.url('/inaccessible-contrast.html'));
     Terra.should.beAccessible({ viewports, rules: ignoredA11y });
+    Terra.it.isAccessible({ viewports, rules: ignoredA11y });
+    it('is accessible', () => {
+      Terra.validates.accessibility({ viewports, rules: ignoredA11y });
+    });
   });
 });