+++ tests/wdio/resize-spec.js
@@ -5,8 +5,8 @@ Terra.viewports('tiny', 'small', 'large').forEach((viewport) => {
       browser.url('/compare.html');
     });
 
-    Terra.should.beAccessible();
-    Terra.should.matchScreenshot();
+    Terra.it.isAccessible();
+    Terra.it.matchesScreenshot();
     it(`resizes ${viewport.name}`, () => {
       const size = browser.getViewportSize();
       expect(size.height).to.equal(viewport.height);